user 곂칠때 막음, 등록된 상태일때 nick 명령어 추가, 인증 관련 함수들 다른 파일로 뺌
