+++ Makefile
@@ -14,7 +14,8 @@ SRCS = ./srcs/main.cpp\
 	   ./srcs/command/comChannelCommand.cpp\
 	   ./srcs/command/comQuit.cpp\
 	   ./srcs/command/comKick.cpp\
-	   ./srcs/command/Bot.cpp
+	   ./srcs/command/Bot.cpp\
+	   ./srcs/command/authenticate.cpp
 OBJS = $(SRCS:.cpp=.o)
 C++ = c++
 FLAGS = -Wall -Wextra -Werror -std=c++98 -I. -g3
